requirements.txt
ccxt==1.49.92
numpy==1.20.3
pandas==1.2.4
pip==21.1.1
scikit-learn==0.24.2
pycoingecko==2.0.0
pathlib
PyYaml
schedule
python-telegram-bot
